update leddriver to use modified library
This was a huge pain. I found a pretty good seeming library
for the PCA9685. But it used a nondefault constructor which
made it impossible to init an array of drivers.
We only need two drivers, but the library is written to be flexible
of course.
So I brought the library over and moved the constructor code into the
begin function.
It compiles now, but still doesn't actually work :(