Clock init and define protections for seed variant
Overriding the generic clock init with the seed init.
The codec _still_ doesnt work unfortunately.
HSE_VALUE triggers a redefined error, but its not the fix even if you mess with the original.
Although it may be part of it, and the original is inside a !defined.
Also added some #ifndef to stop from using the patch_sm when the seed variant is selected.