Seed 1.1 (#37)
* Add rough outline of patch_sm boardsuport

* Flesh out pin numbers

TODO: analog pin names for AnalogRead/Write

* Add analog pinnames to defines

* Add modified pcm3060 codec

Modified to use TwoWire rather than libdaisy i2c.
Patch SM audio still not working...

* Rough out wire use for patch_sm codec

* fiddle with i2c some more

* Fix codec bitshift bug

* Get rid of mid read stop, fix ambiguous warning

* Clock init and define protections for seed variant

Overriding the generic clock init with the seed init.
The codec _still_ doesnt work unfortunately.
HSE_VALUE triggers a redefined error, but its not the fix even if you mess with the original.
Although it may be part of it, and the original is inside a !defined.
Also added some #ifndef to stop from using the patch_sm when the seed variant is selected.

* Add defines for HSE_VALUE and LED_BUILTIN

* Add other periphs to clock init

* Remove commented out HSE define

* Add pin keytwords

* Add board version check function

* Add audio, codec inits, move check to audioclass

* Adjust for patch sm variant

* Revise for more generic patch sm variant

* Move wire setup back to codec

For some reason it doesnt actually work here if you set up
the TwoWire and pass the pointer. Not sure why but it works this way so...

* use seed safe pinnames for codec

* Add boards to keywords

* Fix channel swap bug

* Switch around CV5-8 pin defines

CV5 A9 -> A11
CV6 A12 -> A10
CV7 A10 -> A9
CV8 A11 -> A12

* Add AudioSettings example, add IN OUT defines

* CV Output example

* Port gate in

* Port Gate Output

* Port momentary switch

* Port toggle switch

* Port pot input

* Port CV Input

* Create conversion functions, update CV_Input to use them

* Update Pot Input to use voltage conversion function

* Add WriteCvOut, use in CV_Output example

* Change SM analog pin defines to new format

Figured out how to order them A0, A1 in the variant, just carrying that
through to the SM defines

* Fix swapped pin. PG11 should be PD10

* Use PD3 for version check

This requires adding that pin to the seed variant.